Add files via upload
In order to obtain maximum information from the data, you need to clean it properly. Data comes in a dirty form hence, it requires cleaning. Improperly cleaned data can give a misleading information which will affect decision making. Knowing how to clean a data is an important skill that every analyst should have. The data attached is a way to fill the missing values, change column name and check the type of information on the dataframe.